Update GUI Email Tool.py
Made it so that program can be compiled without needing to connect to intended network.
Clicking the "Connect to Active Directory" button, however, will throw an error if not in the intended domain
All buttons are functional except for "Send Emails", still deciding if I want to implement this in a separate file or not. That being said, final functionality should be coming soon then will tidy up UX as requested but user(s).